feat: move TimeConfig helpers to separate file
