+++ src/server/api/routers/activityDefinitions.ts
@@ -3,7 +3,6 @@ import { createTRPCRouter, protectedProcedure } from '~/server/api/trpc';
 import {
     type ActivityCompletions,
     type ActivityDefinition,
-    computeActivityCompletionsDiff,
     type EndConfig,
     type RepeatConfig,
 } from '~/components/activity/activity-definition-models';
@@ -14,7 +13,10 @@ import {
 } from '~/components/activity/models';
 import { ActivityType as PrismaActivityEnum, Prisma } from '@prisma/client';
 import { type ZoneInfo } from '~/components/zone/models';
-import { type TimeConfig as InternalTimeConfig } from '~/components/time_picker/date';
+import {
+    convertIntToTimeConfig,
+    convertTimeConfigToInt,
+} from '~/components/time_picker/models';
 
 export const RepeatConfigZodSchema = z.object({
     every: z.number(),
@@ -156,57 +158,6 @@ export const ActivitySchema = z.object({
         }),
     ]),
 });
-
-export function convertIntToTimeConfig(time: number): InternalTimeConfig {
-    if (time >= 525600) {
-        return {
-            unit: 'year',
-            value: time / 525600,
-        };
-    } else if (time >= 43200) {
-        return {
-            unit: 'month',
-            value: time / 43200,
-        };
-    } else if (time >= 10080) {
-        return {
-            unit: 'week',
-            value: time / 10080,
-        };
-    } else if (time >= 1440) {
-        return {
-            unit: 'day',
-            value: time / 1440,
-        };
-    } else if (time >= 60) {
-        return {
-            unit: 'hour',
-            value: time / 60,
-        };
-    } else {
-        return {
-            unit: 'minute',
-            value: time,
-        };
-    }
-}
-
-export function convertTimeConfigToInt(timeConfig: InternalTimeConfig): number {
-    switch (timeConfig.unit) {
-        case 'minute':
-            return timeConfig.value;
-        case 'hour':
-            return timeConfig.value * 60;
-        case 'day':
-            return timeConfig.value * 24 * 60;
-        case 'week':
-            return timeConfig.value * 7 * 24 * 60;
-        case 'month':
-            return timeConfig.value * 30 * 24 * 60;
-        case 'year':
-            return timeConfig.value * 365 * 24 * 60;
-    }
-}
 
 const activityWithDetails = Prisma.validator<Prisma.ActivityArgs>()({
     include: {
